+++ src/neuroconv/datainterfaces/__init__.py
@@ -1,5 +1,6 @@
 # Behavior
 from .behavior.audio.audiointerface import AudioInterface
+from .behavior.dannce.danncedatainterface import DANNCEInterface
 from .behavior.deeplabcut.deeplabcutdatainterface import DeepLabCutInterface
 from .behavior.fictrac.fictracdatainterface import FicTracDataInterface
 from .behavior.lightningpose.lightningposedatainterface import (
@@ -184,6 +185,7 @@ interface_list = [
     ExternalVideoInterface,
     InternalVideoInterface,
     AudioInterface,
+    DANNCEInterface,
     DeepLabCutInterface,
     SLEAPInterface,
     MiniscopeBehaviorInterface,
@@ -231,6 +233,7 @@ interfaces_by_category = dict(
     behavior=dict(
         ExternalVideo=ExternalVideoInterface,
         InternalVideo=InternalVideoInterface,
+        DANNCE=DANNCEInterface,
         DeepLabCut=DeepLabCutInterface,
         SLEAP=SLEAPInterface,
         FicTrac=FicTracDataInterface,